+++ src/danae/ARX_Player.cpp
@@ -1804,7 +1804,7 @@ void ARX_PLAYER_LoadHeroAnimsAndMesh()
 
 	if (PAK_FileExist(texscript))
 	{
-		long FileSize = 0;
+		size_t FileSize = 0;
 		io->script.data = (char *)PAK_FileLoadMalloc(texscript, &FileSize);
 
 		if (io->script.data != NULL)
